Move interfaces to common core folder + use dataclass
