+++ src/home_robot/home_robot/core/abstract_agent.py
@@ -1,7 +1,7 @@
 from typing import Dict, Any, Tuple
 from abc import ABC, abstractmethod
 
-from home_robot.core_interfaces import Action, Observations
+from .interfaces import Action, Observations
 
 
 class Agent(ABC):
