+++ src/home_robot/home_robot/core/abstract_perception.py
@@ -1,7 +1,7 @@
 from typing import Any
 from abc import ABC, abstractmethod
 
-from home_robot.core_interfaces import Observations
+from .interfaces import Observations
 
 
 class PerceptionModule(ABC):
